added a timer to game and added scores file stuff
scores file is now read in and sorted (hopefully ?) and displayed (hopefully ?)
a timer was also added to the main game so that we can send that as players score. player will need to input their name and their score and name then sent to file when the game is complete.